requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv==0.12.0
webcolors==1.11.1
requests==2.23.0
asyncio==3.4.3
setuptools==46.1.3
pywikibot==4.1.0
splinter==0.13.0
markovify==0.8.0
schedule==0.6.0
wikitextparser==0.44.0
tabulate==0.8.7
numpy==1.21.2
matplotlib==3.3.3
dateparser==0.7.6
wikipedia==1.4.0
aioify==0.3.2
gspread-asyncio==1.1.0
aiohttp==3.6.2
aiodns==2.0.0
pytz==2020.4
pymongo<5,>=4.1
pymongo[srv]<5,>=4.1
motor==3.0.0
